+++ src/migrations/Install.php
@@ -14,6 +14,7 @@ class Install extends Migration
             'transportSettings' => $this->longText()->null(),
             'transportClass' => $this->string()->notNull(),
             'sent' => $this->integer()->defaultValue(0),
+            'testSuccess' => $this->boolean()->defaultValue(false),
             'ranking' => $this->integer()->defaultValue(0),
             // defaults
             'dateCreated' => $this->dateTime()->notNull(),