+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Craft CMS - Mailer chain
 
+## v1.0.0-beta3 - 2023-02-19
+
+### Changed
+
+- Added requirement to test adapters in the chain before using
+- Added message to the adapter settings view when there is no or not successful tested adapter in the chain
+- Removed custom CP section and added redirect for plugin settings
+- Fixed saving transport settings in JSON format for non-json supporting databases
+
 ## v1.0.0-beta2 - 2023-02-18
 
 ### Changed